Fix serialization breakage due to sympy.Rational now being a numbers.Number - Reorder the sympy checks to come before generic number checks - Rename json.py to json_serialization.py to avoid collisions with the built-in json library
