+++ check/pytest-changed-files
@@ -64,7 +64,7 @@ changed=$(git diff --name-only ${rev} -- \
 if git diff --name-only "${rev}" -- | grep "__init__\.py$" > /dev/null; then
   # Include global API tests when an __init__ file is touched.
   changed+=('docs/docs_coverage_test.py')
-  changed+=('cirq/protocols/json_test.py')
+  changed+=('cirq/protocols/json_serialization_test.py')
 fi
 num_changed=$(echo -e "${changed[@]}" | wc -w)
