+++ dev_tools/bash_scripts_test.py
@@ -150,7 +150,7 @@ def test_pytest_changed_files_file_selection(tmpdir_factory):
                  'echo x > __init__.py\n')
     assert result.exit_code == 0
     assert result.out == ('INTERCEPTED pytest docs/docs_coverage_test.py '
-                          'cirq/protocols/json_test.py\n')
+                          'cirq/protocols/json_serialization_test.py\n')
     assert result.err.split() == (
         "Comparing against revision 'HEAD'.\n"
         "Found 2 test files associated with changes.\n").split()
